пинаю фронт обрабатываю кнопки в заявках, удаление сделано, добавление сделано, закрытие
надо добавить добавление серийника
реализовать закрытие инженером, подключить websocket

TODO
ДОбавить отображение по контрагентам (фронт) и датам(БД)

Вынести константы
Добавление контрагентов с кнопки + возможность добавления своих колонок (не приоритетно);

Добавить списание / поступление оборудования на основе заявки;
Кнопка ручного добавления оборудования на склад;
Кнопка ручного добавления заявки

Добавить Акты(обдумать)
В отображении заявок добавить Печать актов
Вывод окна ошибки (не приоритетно)
Логи (скорее всего во время бетта теста для большего понимания что нужно)
Нормально отловить ошибки(скорее всего на бетте)
Подумать по фронту (см https://htmx.org/docs/)
не забыть путь к файлу акта

Генерация ШК на основе номера заявки;
Прикрутить бот;
Прикрутить почтовый сервер;
Прикрутить АПИ jira;
Тесты (не приоритетно / опционально)

+++ src/main/java/com/javarush/borisov/db/Repository/ContragentRepo.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ContragentRepo extends JpaRepository<Contragent, Integer> {
+public interface ContragentRepo extends JpaRepository<Contragent, Long> {
 
     @Override
     @NonNull
