+++ src/main/java/com/javarush/borisov/db/Repository/UserRepo.java
@@ -7,9 +7,11 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepo extends JpaRepository<User, Integer> {
+public interface UserRepo extends JpaRepository<User, Long> {
 
      Optional<User> findByMail(String email);
+
+
 
 
 }
